moved incoming message details to model module
